codecs: fix pts trancation issue
On the host side: we should use 8 bytes
to pass around pts, instead of 4.

BUG: 168593698

+++ android/android-emu/android/emulation/H264PingInfoParser.cpp
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #include "android/emulation/H264PingInfoParser.h"
-
 #define H264_DEBUG 0
 
 #if H264_DEBUG
@@ -118,7 +117,7 @@ void H264PingInfoParser::parseGetImageParams(void* ptr, GetImageParam& param) {
     param.pDecoderErrorCode = (int*)(retptr);
     param.pRetWidth = (uint32_t*)(retptr + 8);
     param.pRetHeight = (uint32_t*)(retptr + 16);
-    param.pRetPts = (uint32_t*)(retptr + 24);
+    param.pRetPts = (uint64_t*)(retptr + 24);
     param.pRetColorPrimaries = (uint32_t*)(retptr + 32);
     param.pRetColorRange = (uint32_t*)(retptr + 40);
     param.pRetColorTransfer = (uint32_t*)(retptr + 48);
